itw itself warning cleanup: fixed rawtypes and  unchecks, added braces and  Override
After this clean up only "internal proprietary API and may be removed in a future release" warnings remain fro make check. Please keep itw in this way :)

remaining issues:

    icedtea-web/netx/net/sourceforge/jnlp/cache/ResourceTracker.java:357: warning: [deprecation] toURL() in File has been deprecated
    icedtea-web/netx/net/sourceforge/jnlp/cache/CacheUtil.java:128: warning: [deprecation] toURL() in File has been deprecated
    icedtea-web/netx/net/sourceforge/jnlp/runtime/Boot.java:261: warning: [deprecation] toURL() in File has been deprecated
  There have been a lot of work around cach x file x url escaping, and as main difference between file.tourl and file.touri.tourl is escapin, I rather left it.

    icedtea-web/plugin/icedteanp/java/sun/applet/PluginAppletViewer.java:804: warning: [deprecation] Ref in sun.misc has been deprecated
    icedtea-web/plugin/icedteanp/java/sun/applet/PluginAppletViewer.java:121: warning: [deprecation] Ref in sun.misc has been deprecated
  What can be done?

    icedtea-web/netx/net/sourceforge/jnlp/browser/BrowserAwareProxySelector.java:205: warning: [fallthrough] possible fall-through into case
    icedtea-web/netx/net/sourceforge/jnlp/browser/BrowserAwareProxySelector.java:211: warning: [fallthrough] possible fall-through into case
  This have to be fixed

+++ netx/net/sourceforge/jnlp/splashscreen/impls/defaultsplashscreen2012/NatCubic.java
@@ -115,7 +115,7 @@ public class NatCubic extends ControlCurve {
         Cubic[] Y = calcNaturalCubic(pts.npoints - 1, pts.ypoints);
         /* very crude technique - just break each segment up into steps lines */
         Polygon p = new Polygon();
-        p.addPoint((int) Math.round(X[0].eval(0)), (int) Math.round(Y[0].eval(0)));
+        p.addPoint(Math.round(X[0].eval(0)), Math.round(Y[0].eval(0)));
         for (int i = 0; i < X.length; i++) {
             for (int j = 1; j <= STEPS; j++) {
                 float u = j / (float) STEPS;
